made ColorPicker component private and added its sampler in qa-stories
